# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CRUD Barang",
        "description": "This is a sample crud barang with mongodb",
        "contact": {
            "name": "Ahmad Fauzi",
            "email": "fauzi442097@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "description": "Get all users in realtime database",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": "200"
                                        },
                                        "message": {
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "key": {
                                                        "description": "Key ",
                                                        "type": "string",
                                                        "example": "-N61YMhcGqkNze0_tYh6"
                                                    },
                                                    "name": {
                                                        "description": "name user",
                                                        "type": "string",
                                                        "minLength": 3,
                                                        "example": "User Test3"
                                                    },
                                                    "address": {
                                                        "description": "address user",
                                                        "type": "string",
                                                        "example": "Surabaya"
                                                    },
                                                    "email": {
                                                        "description": "email user",
                                                        "type": "string",
                                                        "example": "usertest3@gmail.com"
                                                    },
                                                    "phoneNo": {
                                                        "description": "Phone no",
                                                        "type": "number",
                                                        "minLength": 8,
                                                        "example": "08986162100"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "key": "-N61YMhcGqkNze0_tYh6",
                                                    "name": "User Test3",
                                                    "address": "Surabaya",
                                                    "email": "usertest3@gmail.com",
                                                    "phoneNo": "08986162100"
                                                },
                                                {
                                                    "key": "-N61YK8mgZg2xuOmKSyR",
                                                    "name": "User Test2",
                                                    "address": "Jakarta",
                                                    "email": "usertest2@gmail.com",
                                                    "phoneNo": "08986162400"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create new user",
                "description": "Create new data user to database",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "email",
                                    "phoneNo"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama user",
                                        "type": "string",
                                        "minLength": 3,
                                        "example": "User Test3"
                                    },
                                    "address": {
                                        "description": "address user",
                                        "type": "string",
                                        "example": "Surabaya"
                                    },
                                    "email": {
                                        "description": "email user",
                                        "type": "string",
                                        "example": "usertest3@gmail.com"
                                    },
                                    "phoneNo": {
                                        "description": "phone no user",
                                        "type": "number",
                                        "minLength": 1,
                                        "example": "08986162100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Created"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "description": "Nama user",
                                                    "type": "string",
                                                    "minLength": 3,
                                                    "example": "User Test3"
                                                },
                                                "address": {
                                                    "description": "address user",
                                                    "type": "string",
                                                    "example": "Surabaya"
                                                },
                                                "email": {
                                                    "description": "email user",
                                                    "type": "string",
                                                    "example": "usertest3@gmail.com"
                                                },
                                                "phoneNo": {
                                                    "description": "phone no user",
                                                    "type": "number",
                                                    "minLength": 1,
                                                    "example": "08986162100"
                                                },
                                                "key": {
                                                    "description": "key of user",
                                                    "type": "string",
                                                    "example": "-N61YMhcGqkNze0_tYh6"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                },
                                                "address": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The address field is required."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    }
                                                },
                                                "phoneNo": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The phoneNo field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{key}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Find user by key",
                "description": "Find data user in database by Key",
                "operationId": "51d3d3d93274ac53f5fb3a6c95a77f07",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/key"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": "200"
                                        },
                                        "message": {
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "key": {
                                                    "description": "Key ",
                                                    "type": "string",
                                                    "example": "-N61YMhcGqkNze0_tYh6"
                                                },
                                                "name": {
                                                    "description": "name user",
                                                    "type": "string",
                                                    "minLength": 3,
                                                    "example": "User Test3"
                                                },
                                                "address": {
                                                    "description": "address user",
                                                    "type": "string",
                                                    "example": "Surabaya"
                                                },
                                                "email": {
                                                    "description": "email user",
                                                    "type": "string",
                                                    "example": "usertest3@gmail.com"
                                                },
                                                "phoneNo": {
                                                    "description": "Phone no",
                                                    "type": "number",
                                                    "minLength": 8,
                                                    "example": "08986162100"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Response Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{key}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete existing user",
                "description": "Delete existing data user in database by Key",
                "operationId": "f6383df382ada2d83954fc94a916b96b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/key"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Success Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": "200"
                                        },
                                        "message": {
                                            "example": "Success deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Response Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "NotFoundResponse": {
                "title": "Response Not Found",
                "properties": {
                    "code": {
                        "type": "number",
                        "example": "404"
                    },
                    "message": {
                        "type": "string",
                        "example": "Data Not Found"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "key": {
                "name": "key",
                "in": "path",
                "description": "Key of data user",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
        }
    }
}